be compatible with newer fysom versions
The waiting_timeout callback now always fires (even when there is
no state transition). In previous versions of fysom this was not the
case.

Since the timeout event with transition pending -> pending (meaning we
already started, so we don't care about the waiting timeout) will invoke
the callback we need to check if the timeout is still relevant.